backend/pyproject.toml
[project]
name = "mailcake-backend"
version = "0.1.0"
description = "MailCake - AI-powered email summarization service"
requires-python = ">=3.11"
dependencies = [
    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",

    # Database
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.0",

    # Redis + Task Queue
    "redis>=5.0.0",
    "arq>=0.26.0",          # async task queue
    "apscheduler>=3.10.0",  # scheduled jobs

    # LLM (OpenAI SDK → LiteLLM Proxy)
    "openai>=1.50.0",

    # Gmail / Email
    "google-auth>=2.35.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.150.0",
    "imapclient>=3.0.1",
    "mail-parser>=3.15.0",

    # Auth & Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.12",
    "cryptography>=43.0.0",

    # Utils
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "aiosmtplib>=3.0.0",  # async email sending
    "jinja2>=3.1.4",       # email templates
    "beautifulsoup4>=4.12.0",  # HTML email parsing
    "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
